requirements.txt
streamlit
Pillow
matplotlib
pytesseract
pandas
openai
gspread
pillow-heif
google-auth
google-auth-oauthlib
